% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewExons.R
\name{addNewExons}
\alias{addNewExons}
\title{addNewExons()}
\usage{
addNewExons(annot_gtf, gene_ids, fasta, canonical = TRUE, seed = 1234)
}
\arguments{
\item{annot_gtf}{table containing gtf info for annotated transcripts}

\item{gene_ids}{genes for which transcript with novel exon should be added}

\item{canonical}{new exons are canonical (GT donor site, AG acceptor site) [default = TRUE]}

\item{seed}{seed for reproducability [default = 1234]}
}
\value{
the updated table containing gtf info for the annotated transcripts and the novel transcripts
}
\description{
generate a new transcript containing a novel exon for each gene in gene_ids
}
