% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewJunctions.R
\name{addNewJunctions}
\alias{addNewJunctions}
\title{addNewJunctions()}
\usage{
addNewJunctions(annot_gtf, gene_ids, seed = 1234)
}
\arguments{
\item{annot_gtf}{table containing gtf info for annotated transcripts}

\item{gene_ids}{genes for which transcript with novel junction should be added}

\item{seed}{seed for reproducability [default = 1234]}
}
\value{
the updated table containing gtf info for the annotated transcripts and the novel transcripts
}
\description{
generate a new transcript containing a novel junction for each gene in gene_ids
}
