% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCharacteristics.R
\name{DataCharacteristics}
\alias{DataCharacteristics}
\title{DataCharacteristics}
\usage{
DataCharacteristics(inDir)
}
\arguments{
\item{inDir}{directory containing output from characterizeData}
}
\value{
object of class "DataCharacteristics"
}
\description{
data characteristics based on existing Smart-seq3 data
}
