% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRandomClusterPSI.R
\name{getRandomClusterPSI}
\alias{getRandomClusterPSI}
\title{getRandomClusterPSI()}
\usage{
getRandomClusterPSI(gtf, nClusters, fixedGenes, diffGenes, seed)
}
\arguments{
\item{gtf}{table containing gtf info for all transcripts}

\item{nClusters}{number of clusters}

\item{fixedGenes}{genes that have the same PSI values across clusters}

\item{diffGenes}{genes that have different PSI values across clusters}

\item{seed}{for reproducibility [default = 1234]}
}
\value{
matrix containing PSI value for each cluster and transcript
}
\description{
generate random cluster PSI values
}
