% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRandomGeneExpression.R
\name{getRandomGeneExpression}
\alias{getRandomGeneExpression}
\title{getRandomGeneExpression()}
\usage{
getRandomGeneExpression(
  nClusters,
  nCellsPerCluster,
  genes,
  dataChar,
  minCount = 0,
  seed = 1234
)
}
\arguments{
\item{nClusters}{number of clusters}

\item{nCellsPerCluster}{number of cells per cluster}

\item{genes}{geneIDs to generate read counts for}

\item{dataChar}{data characteristics object}

\item{minCount}{minimum gene expression [default = 0]}

\item{seed}{for reproducability [default = 1234]}
}
\value{
matrix containing gene expression for each cell
}
\description{
generate random gene expression based on gene expression from data.
  gene expression for cells within the same cluster is only different by cellSize
}
