% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTranscripts.R
\name{getTranscripts}
\alias{getTranscripts}
\title{getTranscripts()}
\usage{
getTranscripts(
  gtf,
  fasta,
  nTx,
  out,
  genes = NULL,
  nRandomGenes = NULL,
  minTxLen = 1000,
  seed = 1234
)
}
\arguments{
\item{gtf}{gtfFile containing the reference annotation}

\item{fasta}{fastaFile containing the reference genome}

\item{nTx}{number of transcripts per gene}

\item{out}{outdir to write geneInfo file}

\item{genes}{list of genes}

\item{nRandomGenes}{number of random genes for which transcripts are selected (not used if list of genes given)}

\item{minTxLen}{minimum transcript length [default = 1000]}

\item{seed}{seed for reproducability [default = 1234]}
}
\value{
gtf table containing all selected genes and transcripts
}
\description{
get transcripts to simulate
}
