% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCharacteristics.R
\name{sampleBCQual}
\alias{sampleBCQual}
\title{sampleBCQual}
\usage{
sampleBCQual(dataChar, rLen, n)
}
\arguments{
\item{dataChar}{DataCharacteristics object}

\item{rLen}{length of barcode}

\item{n}{number of barcode qualities to sample}
}
\description{
sample barcode quality for barcode of length rLen
}
