% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCharacteristics.R
\name{sampleNonUMIvsUMI}
\alias{sampleNonUMIvsUMI}
\title{sampleNonUMIvsUMI}
\usage{
sampleNonUMIvsUMI(dataChar, n)
}
\arguments{
\item{dataChar}{DataCharacteristics object}

\item{n}{number of elements to sample}
}
\description{
sample n elements from the distribution of number of nonUMI reads vs. UMI reads
}
