% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCharacteristics.R
\name{sampleNumReadsPerUMI}
\alias{sampleNumReadsPerUMI}
\title{sampleNumReadsPerUMI}
\usage{
sampleNumReadsPerUMI(dataChar, n)
}
\arguments{
\item{dataChar}{DataCharacteristics object}

\item{n}{number of elements to sample}
}
\description{
sample n elements from the distribution of number of reads per UMI
}
