% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCharacteristics.R
\name{sampleReadQual}
\alias{sampleReadQual}
\title{sampleReadQual}
\usage{
sampleReadQual(dataChar, rLen, n)
}
\arguments{
\item{dataChar}{DataCharacteristics object}

\item{rLen}{read length}

\item{n}{number of read qualities to sample}
}
\description{
sample read quality for read of length rLen
}
