% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCharacteristics.R
\name{sampleUMIcounts}
\alias{sampleUMIcounts}
\title{sampleUMIcounts}
\usage{
sampleUMIcounts(dataChar, n, minCount = 0)
}
\arguments{
\item{dataChar}{DataCharacteristics object}

\item{n}{number of elements to sample}

\item{minCount}{minimum number of counts}
}
\description{
sample n elements from the distribution of UMIcounts
}
