% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCharacteristics.R
\name{sampleUMIfragLen}
\alias{sampleUMIfragLen}
\title{sampleUMIfragLen}
\usage{
sampleUMIfragLen(dataChar, n, minLen)
}
\arguments{
\item{dataChar}{DataCharacteristics object}

\item{n}{number of elements to sample}

\item{minLen}{minumum fragment length}
}
\description{
sample n elements from the distribution of fragment length for reads containing UMIs
}
