% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateReads.R
\name{simulateReads}
\alias{simulateReads}
\title{simulateReads()}
\usage{
simulateReads(
  transcriptExpression,
  transcript_seq,
  dataChar,
  out,
  barcodeLen = 8,
  UMIlen = 10,
  readLen = 150,
  tag = "ATTGCGCAATG",
  PAF = 2,
  errorRate = 0.005,
  seed = 1234
)
}
\arguments{
\item{transcriptExpression}{expression for each transcript and cell}

\item{transcript_seq}{sequence for each transcript}

\item{dataChar}{data characteristics}

\item{out}{output directory to write reads to}

\item{barcodeLen}{length of barcode [default = 8]}

\item{UMIlen}{length of UMI [default = 10]}

\item{readLen}{length of read [default = 150]}

\item{tag}{tag to indicate UMI containing read [default = "ATTGCGCAATG"]}

\item{PAF}{PCR amplification factor [default = 2]}

\item{errorRate}{errorRate per bp [default = 0.005]}

\item{seed}{for reporducibility [default = 1234]}
}
\description{
simulate Smart-seq3 reads for selected transcripts
}
