% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTrueTranscripts.R
\name{writeTrueTranscripts}
\alias{writeTrueTranscripts}
\title{writeTrueTranscripts()}
\usage{
writeTrueTranscripts(gtf, fasta, out)
}
\arguments{
\item{gtf}{gtf table containing ground truth transcripts}

\item{fasta}{fastaFile containing the reference genome}

\item{out}{outdir to write resulting gtf and fasta file}
}
\value{
transcript sequences as DNASTringSet object
}
\description{
write ground truth to files (gtf annotation and fasta with sequence for each transcript)
}
